code/kafka/producer/requirements.txt
confluent_kafka
faker
deep-translator
python-dotenv